# swagger.yaml
openapi: 3.0.3
info:
  title: SuperFrigo
  description: API description
  version: 1.0.0
servers:
  - url: https://back-end-production-d316.up.railway.app
components:
  schemas:
    Ingredient:
      type: object
      properties:
        name:
          type: string
          #default: mozzarella
        expiration:
          type: string
          #default: 01/01/1970
        quantity:
          type: integer
          format: int32
          #default: 100
    Recipe:
      type: object
      properties:
        title:
          type: string
        ingredients:
          type: array
          items:
            $ref: '#components/schemas/Ingredient'
        image:
          type: string
        link:
          type: string
        data:
          type: array
          items:
            type: object

paths:
  /login:
    post:
      summary: Login api.
      description: Used to log in
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                email:
                  type: string
                  example: "esempio@gmail.com"
                psw:
                  type: string
                  example: "HashedPassword"
      responses:
        '200':    # status code
          description: SID (session id)
          content:
            json:
              schema: 
                type: string
        '400':    # status code
          description: Error
          content:
            json:
              schema: 
                type: string
                example: Wrong Credentials
  /register:
    post:
      summary: Register api.
      description: Used to register
      #parameters:
      #  - name: mail
      #    in: query
      #    description: email to register
      #    required: true
      #    schema:
      #      type: string
      #      default: user@email.com
      #  - name: psw
      #    in: query
      #    description: hashed psw to register
      #    required: true
      #    schema:
      #      type: string
      #      default: HashedPassword
      #  - name: username
      #    in: query
      #    description: the username
      #    required: true
      #    schema:
      #      type: string
      #      default: username
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                mail:
                  type: string
                psw:
                  type: string
                username:
                  type: string
              example:   # Sample object
                mail: user@email.com
                username: example_username
                psw: HashedPassword
      responses:
        '200':    # status code
          description: Ok code
          content:
            json:
              schema: 
                type: string
                example: ok
        '400':    # status code
          description: Error
          content:
            json:
              schema: 
                type: string
                example: email already used
        '401':    # status code
          description: Error
          content:
            json:
              schema: 
                type: string
                example: email not valid
  /add:
      post:
        summary: Add ingredient
        description: Add an ingredient to the user personal storage
        #parameters:
        #  - name: sid
        #    in: query
        #    description: SID
        #    required: true
        #    schema:
        #      type: string
        #      default: 000SessionID000
        #  - name: ingredient
        #    in: query
        #    description: The ingredient to add
        #    required: true
        #    schema:
        #      $ref: '#components/schemas/Ingredient'
        requestBody:
          content:
            application/json:
              schema:      # Request body contents
                type: object
                properties:
                  sid:
                    type: string
                  ingredient:
                    $ref: '#components/schemas/Ingredient'
                example:   # Sample object
                  sid: 000SessionID000
                  ingredient:
                    name: Mozzarella
                    expiration: 01/01/1970
                    quantity: 100
        responses:
          '200':    # status code
            description: Ok code
            content:
              json:
                schema: 
                  type: string
                  example: ok
          '401':    # status code
            description: Error
            content:
              json:
                schema: 
                  type: string
                  example: bad ingredient format
          '403':    # status code
            description: Error
            content:
              json:
                schema: 
                  type: string
                  example: session not found
  /ingredients:
      get:
        summary: list of ingredients of the user
        description: Get all the ingredients in the user storage
        parameters:
          - name: sid
            in: query
            description: SID
            required: true
            schema:
              type: string
              default: 000SessionID000
        responses:
          '200':    # status code
            description: List of ingredients
            content:
              json:
                schema: 
                  type: array
                  items:
                    $ref: '#components/schemas/Ingredient'
          '403':    # status code
            description: Session not found
            content:
              json:
                schema: 
                  type: string
                  example: session not found
  /all_ingredients:
      get:
        summary: All ingredients
        description: list of all ingredients in the DB
        responses:
          '200':    # status code
            description: List of ingredients
            content:
              json:
                schema: 
                  type: array
                  items:
                    type: object
  /recipes:
      get:
        summary: Get possible recipes
        description: Get all the possible recipes with the user storage, ordered by number of missing ingredients (ascending order)
        parameters:
          - name: sid
            in: query
            description: SID
            required: true
            schema:
              type: string
              default: 000SessionID000
        responses:
          '200':    # status code
            description: List of recipes
            content:
              json:
                schema: 
                  type: array
                  items:
                    $ref: '#components/schemas/Recipe'
          '403':    # status code
            description: Session not found
            content:
              json:
                schema: 
                  type: string
                  example: session not found
  /guest_registration:
      get:
        summary: Login for guests
        description: Generate an unique id for the guest
        responses:
          '200':    # status code
            description: ID
            content:
              json:
                schema: 
                  type: integer
                  example: 1
  /user:
      get:
        summary: Return the username from a sid
        parameters:
          - name: sid
            in: query
            description: SID
            required: true
            schema:
              type: string
              default: 000SessionID000
        responses:
          '200':    # status code
            description: Error
            content:
              json:
                schema: 
                  type: object
                  properties:
                    guest:
                      type: boolean
                      example: true
                    username:
                      type: string
                      example: "guest"
          '403':    # status code
            description: Error
            content:
              json:
                schema: 
                  type: string
                  example: session not found   
          '400':    # status code
            description: Error
            content:
              json:
                schema: 
                  type: string
                  example: user not found

  /change_password:
    post:
      summary: Change the user password
      description: The new and the old password must be hashed
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                sid:
                  type: string
                old_psw:
                  type: string
                new_psw:
                  type: string
              example:   # Sample object
                sid: 000SessionID000
                old_psw: Password1
                new_psw: Password2
      responses:
        '200':    # status code
          description: Ok code
          content:
            json:
              schema: 
                type: string
                example: ok
        '402':    # status code
          description: Error
          content:
            json:
              schema: 
                type: string
                example: the new password is the same as the old one
        '403':    # status code
          description: Error
          content:
            json:
              schema: 
                type: string
                example: session not found
        '405':    # status code
          description: Error
          content:
            json:
              schema: 
                type: string
                example: old password does not correspond
        '404':    # status code
          description: Error
          content:
            json:
              schema: 
                type: string
                example: user not found
  /remove:
    delete:
      summary: Remove an ingredient from the user storage, given the index from the storage
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                sid:
                  type: string
                ingredient:
                  type: string
              example:   # Sample object
                sid: 000SessionID000
                ingredient: 0
      responses:
        '200':    # status code
          description: Ok code
          content:
            json:
              schema: 
                type: string
                example: ok
        '403':    # status code
          description: Error
          content:
            json:
              schema: 
                type: string
                example: session not found
        '402':    # status code
          description: Error
          content:
            json:
              schema: 
                type: string
                example: invalid index
        '400':    # status code
          description: raised when user never added an ingredient
          content:
            json:
              schema: 
                type: string
                example: storage not found
        '401':    # status code
          description: Error
          content:
            json:
              schema: 
                type: string
                example: generic error
  /delete_account:
    delete:
      summary: Delete the user
      description: Used to delete all data of an account
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                sid:
                  type: string
              example:   # Sample object
                sid: 000SessionID000
      responses:
        '200':    # status code
          description: Ok code
          content:
            json:
              schema: 
                type: string
                example: ok
        '403':    # status code
          description: Error
          content:
            json:
              schema: 
                type: string
                example: session not found
        '404':    # status code
          description: Error
          content:
            json:
              schema: 
                type: string
                example: user not found